Fix cursor goes to left margin
Use case. Reading a line when the cursor is not at the left margin without prompt (System.out.print something): the user has written something and then deletes the input or moves the cursor back to the first position, the cursor instead of going to the starting column it goes back all the way to the left margin.